Detect non-ascii hosts and refuse to build on them
The htparse library makes some assumptions about characters that are
not true for non-ascii-based system.  (For example, that all
upper-case letters are numerically between 'A' and 'Z'; or that
(int)'A' is the same as the encoding of 'h' on the wire.)  This is
always true in practice, except for some really horrible places we
will never want to build. But in theory, C allows (int)'A' to be
basically anything.  So let's put our pedantic hats on briefly and
detect non-ASCII environments, if only to give an error.