Add ability for rules to be ordered.
Prior to this patch callback matches were as follows:

1. check for hash match
2. check for glob match
3. check for regex match

This can lead to confusion when creating a rule structure for uris. For example
if you have:

regex /stuff/[0-1]+/
glob  /stuff/*

The glob will always match and never get to the regex. Now rules are processed
as they were defined in order. So if in the above situation you would do:

evhtp_set_regex_cb(..., "/stuff/[0-1]+/", ...);
evhtp_set_glob_cb(..., "/stuff/*", ...);

And the regex is assured to be called before the glob.