Updates for threading functionality.
[general]

All of the thread-new-type functions will be deprecated in lieu of @jgli's
additions, which allow for entry and exit callbacks. But not just yet.

Reviewing the pull requst for the above, the function declarations had been
changed without a warning to the user. So this commit attempts to keep backwards
compatibility, but will also warn the user of the upcoming changes.

Once we feel as if these minor updates have been ported by the majority of
users, we rename (or deprecate) the 'wexit' functions to their respective
names.

[evhtp-config]
- Added cross-compile macros for marking a function declaration as deprecated.
  It's been cut and pasted to so many projects, hell if I know who wrote it or
  under what license.

[evthr]

- Moved `evthr_[pool_]new` to `static _evthr_[pool_]new`; which takes both the entry and exit
  callbacks.

- The exported function `evthr_[pool_]new` now calls `_evthr_[pool_]new` but with a NULL exit
  callback.

- The exported function `evthr_[pool_]wexit_new` calls `_evthr_[pool_]new with the new exit
  callback.

- These two functions are to be removed once merged with their parents which are
  now marked as deprecated

[evhtp]

- The same logic applies to the local wrapper functions in evhtp.c
    `evhtp_use_threads` is now `static _evhtp_use_threads`, taking both the
    and exit callbacks. (which in turn; uses `evthr_pool_wexit_new`)

    `evhtp_use_threads` now calls `_evhtp_use_threads` with the exit callback
     set to NULL, and vice-versa for 	`evhtp_us_threads_wexit`

- While these changes can go mostly unnoticed, the public function
  `evhtp_use_threads` has been marked as deprecated, and will take
   on the syntax of `evhtp_use_threads_wexit`