[Issue#20] Include pathing
As stated within https://github.com/criticalstack/libevhtp/issues/20,
headers were being installed to include/evhtp/* but were not reflected
in the source.

This means that even though evhtp specific include files were in
/usr/include/evhtp/blah.h, a user still had to use -I/usr/include/evhtp
which is just dumb.

This is a holdover until we do a proper code refactor that splits up
various things logically.

So for now, ./include/evhtp == public
            ./include/. == private

Aside from the include/evhtp.h which is a placeholder that points
to include/evhtp/evhtp.h