Remove built-in Oniguruma, now rely on system only
@flokli (github) informed us of our failure to keep up with security
updates in the builtin version of onigurama we ship with evhtp.

- CVE-2017-9224
- CVE-2017-9225
- CVE-2017-9226
- CVE-2017-9227
- CVE-2017-9228
- CVE-2017-9229

We should have never done this in the first place, but we did, and for
that we apologize.

So we decided to completely remove the packaged version. Sorry, but
we think that this is a good thing. Instead, cmake will attempt to
find a system-wide installed version of this library, and if it fails,
regex support will be disabled.

* DELETED ALL BUILTIN ONIGURUMA FILES
* Added FindOniguruma.cmake for the obvious
* Updated examples/test.c to make sure regex is working (or unsupported)
* Moved various parts of the main CMakeLists.txt into smaller ones