Extensive example of streaming data without hogging memory
This is an example as a reply to
https://github.com/criticalstack/libevhtp/issues/47

examples/example_chunked.c will open up a file (of any size),
and stream data back to the client in 128 byte chunked segments.

This is the proper way of sending large amounts of data back to
a client without consuming much memory.