[#76] keep consistent member type of struct evthr
The evthr struct references `exit_cb` of type `evthr_init_cb`.
The `evthr_init_cb` is a typedef to `void (*)(evthr_t *, void *)`

Both of these typedefs are of the same attributes, so nothing bad
happens here; it's more about readability. This patch simply changes
the value type definition of `exit_cb` to the aliased type
`evthr_exit_cb`.

Nothing changes, nothing was "broken", but we need to be consistent.