[#84] fix for hanging connections via resumption.
- When a evhtp request/connection is paused, then resumed, it
  requires EV_WRITE to be triggered (and whatever underlying
  callbacks). If the connections output buffer is greater than
  zero when resumecb_ is called, it will now manually trigger
  the bufferevent's write event via `bufferevent_trigger_event` for
  libevent versions >= 2.1.5.

QUICK REVIEW OF PAUSE/RESUME:

When you "pause" a request/connection, it sets a flag on the
underlying connection 'EVHTP_CONN_FLAG_PAUSED', then disables
the READ side of the bufferevent.

When you RESUME, the FLAG_PAUSED is unset, and `event_active` is used
to call the resume event (which calls `htp__connection_resumecb_`).

If there is still data in the output buffer of the bufferevent; the
connection is set to `EVHTP_CONN_FLAG_WAITING`, or in english: "waiting
to finish writing". It must be noted that EV_READ is not enabled yet,
not until the write callback has been called. It then calls
`bufferevent_trigger_event` for EV_WRITE, which will eventually
send off all the data and call `htp__connection_writecb_`.

The `htp__connection_writecb_` function will then check to see if the
connection is in the `WAITING` state, and if so, unset it, and enable
EV_READ again on the bufferevent.